sql修改,实体类修改,拉取应用调试成功 王成

+++ BeeEdu/src/main/java/com/enablue/mapper/AppDetailMapper.java
@@ -40,5 +40,5 @@ public interface AppDetailMapper {
      * @param applicationTypeId
      * @param serviceTotal
      */
-      List<AppDetail> queryAppDetailByType(@Param("applicationTypeId") Integer applicationTypeId,  @Param("serviceTotal") Long serviceTotal);
+      List<AppDetail> queryAppDetailByType(@Param("applicationTypeId") Integer applicationTypeId,  @Param("serviceTotal") Integer serviceTotal);
 }
